Task 0-7

+++ functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
 
 int print_last_digit(int n)
 {
-	int abs;
+	long int abs;
 
 	if (n > 0)
 		abs = n;
